`actix-web` integration with builtin server function handler route
